requirements.txt
certifi==2022.6.15
distlib==0.3.5
filelock==3.8.0
install==1.3.5
mysql-connector-python==8.0.30
pipenv==2022.8.15
platformdirs==2.5.2
protobuf==3.20.1
PySide2==5.15.2.1
shiboken2==5.15.2.1
virtualenv==20.16.3
virtualenv-clone==0.5.7
